Lazy AcceleoPackageRegistry and Ecore indexing
This changes the design of the AcceleoPackageRegistry in a substantial
way. The main objective is to avoid loading all the Ecore files which
are in a given workspace even if they are not used by Acceleo in any
way.

This commit adds support for EPackage.Descriptor instances in the
registry. These instances are already used by EMF Core to represent
EPackages which are registered through the extension point.
The actual loading of the EPackage is then done through :
  org.eclipse.emf.ecore.EPackage.Descriptor.getEPackage()
(see EcoreEPackageSAXHandler)

descriptors are created from this and added in the registry.

Acceleo supports subtle use cases like for instance a dynamic Ecore
referencing another dynamic ecore which extends classes from a third
Ecore model which is static (with its generated code)
The logic supporting those is kept but has been moved into :
org.eclipse.acceleo.common.internal.utils.LazyEPackageDescriptor.getEPackage()

Clients of the registry should not notice any difference as any call to
org.eclipse.acceleo.common.internal.utils.AcceleoPackageRegistry.getEPackage(String)
will trigger the loading of the Ecore model and that's standard behavior
in EMF. Yet any client doing that for every possible nsURI defeats the
whole lazyness. AcceleoCompletionProcessor is one of those. In the

AcceleoCompletionProcessor.computeEPackageProposals(List<ICompletionProposal>)

method, it iterates over all the keysets and retrieve the EPackages to
propose the nsURI. This commit changes this method so that
AcceleoCompletionProcessor handles the introduced LazyEPackageDescriptor
which gives him all the required information, not needing any loading
of the Ecore model.

Some clients of the AcceleoPackageRegistry also happens to register
themselves Ecore instances from time to time. I tried to keep that the
way it is.

Regarding WorkspaceListening, when initialized the listener will now
start by browsing the whole workspace and create the corresponding
descriptors. This fix an issue I noticed along the way :
 most use cases involving only http:// nsURIs will not work
unless the Acceleo Editor registered himself the files "by hand".
Somebody launching the generation without launching an Acceleo Editor
first would get a failure.

From the end user point of view the difference of resource consumption
(both CPU and memory) is highly noticable. Even more when the workspace
has many Ecore files or some very big ones (like UML.ecore) which are
not used in any way by a generator.

Measures remains to be done.
Other clients within the Acceleo code itself might need the same kind of
changes which have been done on the AcceleoCompletionProcessor or will
defeat the whole point.

Change-Id: Ieeed3f4223d151a06a5e725e524530ccb5a98be3