[427094] Strenghten the code for "weird" Ecore files
It prevent the EmptyStackException to arise in the case of an Ecore file
having multiple roots. The non-first roots will just be ignored.
It also enforce the code so that any failure at the point of creating a
descriptor does not make the whole calling process fail. The implication is
that 'weird' Ecore files will not fail the loading of
AcceleoCommonPlugin, they will be ignored alltogether from Acceleo.