[462315] lambda do not capture their surrounding environment
 Evaluation of lambda must be done in the environment where the lambda
has been defined. If they're not, they do not capture variable
definitions. 

As a consequence, when a variable selector is defined in the
environment, the following expression returns an empty list even though
the selector variable is defined to 'st' for instance. 

   ['str','str1','no']->select(i | i.startsWith(selector))
   
Fix of this requires using a value for lambda that embodies the lambda
itself and the environment so that the lambda can be evaluated at a
later time with it's defining environment captured. 

Environments are changed from simple Map to scoped map made of a stack
of maps. These scoped environments allow the redefinition of variables
in a structured fashion. 

When a lambdaValue is returned in the evaluator, a copy of the
environment is made so that poping the environment has no impact on the
lambda which could leak if the lambda become a first class value some
time. 


Fixed a problem with antlr version used to build. 

Bug: 462315
Change-Id: I5f59956df9505459ef7c7e5e994d6098f182cca6
Signed-off-by: Romain Guider <romain.guider@obeo.fr>