[465272] Adapt IQueryEvaluationEngine to have a single responsability
Instead of passing a "String" which is getting parsed by a hard coded
IQueryBuilderEngine, get an AstResult as a parameter so that this class
doesn't have the responsability to parse the expression.

Also check for "Nothing" before returning the result to avoid exposing
such an implementation detail on the caller.

Bug: 465272
Change-Id: I955e0d315baa0da08e8fa0d0e31ca1e7e121a569
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>