[465272] Introduce static utilities to create Impl instances
Introduces Query, QueryEvaluation, QueryParsing and QueryValidation as
classes providing utility methods for each aspect of the query
management. These are exposed are part of the API and are creating the
actual implementation instances which are in the
org.eclipse.acceleo.query.runtime.impl package and are only exposing the
interfaces. This gives us a fairly wide margin to evolves the
implementation classes and add/adapt the way instances are getting
created while being easier to keep API compatibility.

Make all the **Type classes I could final so that evolving the
query.validation.type package will still be quite easy.


Bug: 465272
Change-Id: I38a383192b7a1bb12a37590a9c56b09b7c0f9640
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>