[483577] Add getTypes(String) returning a list of matching EClassifiers
Sirius had to rely on IEPackageProvider.getType(String) which got
deprecated for good reasons as it would only return a single type
matching the name or throw an exception. 
This commit introduce getTypes(String) which returns a collection of all
the types known by AQL which are matching the name, so that Sirius can
pass this information around.

Bug: 483577