Rework the doc generator to embed STD services in the general doc
The generator now create a document aggregating all the known services
in a single HTML buffer. This document gets directly integrated "in
place" into index.html instead of referencing elements : this enables
CTRL+F searches in the page and I witnessed several users completely
missing the reference documentation.

This commit adapts the syntax of the generated signature to be more in
line with what the user see when AQL is used, for instance:

  asOrderedSet(collection: java.util.Collection) = Set
  
will now be generated:

  Collection->asOrderedSet() : OrderedSet
  
Also operators are dynamically translated in their prefered form:

  add(set: java.util.Set, collection: java.util.Collection) = Set

become 

  Sequence + Collection : Sequence
  
The old generators is still present and is called when client code use
the old methods.

Javadoc embedded in the services is slightly reworked so that the
examples are using the operator syntax.

Change-Id: I6c8ebbd169809790640ed4871c01a03324dc9f74