[LS]Add editor support for "go to definition"
This feature is enabled in the generic editor when ctrl +
left-clicking on elements (e.g. an Acceleo Query, an Acceleo Template,
the parameter of a Query, the parameter of a Template, a term in an AQL
Expression, etc.).

For now it is available on most elements, both from Acceleo and AQL,
although we could probably support additional cases in the future (e.g.
metamodel declarations, EStructuralFeatures, etc.).
For most elements on which it is available, it works as expected
(although the destination selection could gain to be reduced, i.e. to
select only the header of a Template when going to its definition),
including on imported Acceleo modules.
For other elements, like Java-based
IServices or EMF-based IServices, this first implementation uses
placeholders that we will replace in the future to integrate the
resolution mechanism of Acceleo 4 for accessing external Java/EMF
resources.

* Adds a new API to Acceleo that captures links from an element of the
AST of an Acceleo Module to an element of the AST of another Acceleo
Module: AcceleoLocator.
* Adds a similar API for AQL, links can point to another AQL element or
to some "any" element to be interpreted (Acceleo ASTNode, Java, etc.):
AqlLocator. When necessary, the Acceleo API delegates down to the AQL
API while providing it with the necessary context information (variables
of the context, registered services, ...).
* Exposes a few methods from AQL/Acceleo validation so we can reuse
some of their functionalities like type checking and IService
resolution.
* Implements the transformation from these Acceleo/AQL APIs to the LSP4J
API used for links (AbstractLocationLink => Locationlink).
* Plugs the Acceleo resolution mechanism to the Acceleo Language Server.
* Dependencies to Eclipse core and ui are externalize into
org.eclipse.acceleo.aql.ide.ui.

Known issues:
* When a module is updated, the other modules that depend on it are not
updated so there may be issues for going to definitions of their
elements.

Change-Id: Iacb1b1ec83d6876b1892a9b088c6b344fdbf1412
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>