[MM]Fix Documentation was contained by both DocumentedElement and Module
In the Acceleo metamodel, the Module contains ModuleElements which
include Comments and Documentations found in it. ModuleElements contain
their Documentation (the last Documentation located before the
ModuleElement itself).
Thus, the Documentation of a ModuleElement is supposed to be contained
by both, which is not possible, and so the eOpposite reference from a
Documentation to its DocumentedElement fails.

Now a DocumentedElement *references* its Documentation instead of
containing it.

Change-Id: I4c17286e8c5af97dd4ef30a64703edfd9b3638d6
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>