[LS]Add workspace concept to the Acceleo Language Server.
Previously, each AcceleoTextDocument had its own environment, which
made complex handling several Acceleo Modules with a shared "classpath".
This aims at introducing shared IAcceleoEnvironments between several
AcceleoTextDocuments.

* Add a notion of Acceleo "workspace" and "project" to the LS. An
AcceleoProject is a container of AcceleoTextDocuments which share the
same IAcceleoEnvironment. In the case of the Eclipse client, this easily
maps to the IProjects in the workspace of the client, since our
resolution is based on the JDT. An AcceleoWorkspace is a container of
AcceleoProjects. The Acceleo LS has exactly one AcceleoWorkspace which
contains several AcceleoProjects, each containing several
AcceleoTextDocuments.
* AcceleoWorkspace/AcceleoProject are not tied to a particular client
platform, but are provided through a client-specific implementation of
AcceleoLanguageServerContext. The Eclipse-specific implementation
provided is able to map the running platform's IWorkspace into an
AcceleoWorkspace, and maintains it consistent with it during the
IWorkspace's lifecycle by listening to changes in the Eclipse workspace.
* Add API to IAcceleoEnvironment to enable the "unregistering" of a
module. This is necessary for when a file is deleted from the workspace,
or a file is changed and the services it provided change. The
AcceleoEnvironment implementation now makes sure to unregister any
provided services before re-registering a module.
* When modifying an Acceleo file, errors are displayed in the file as
the user modifies it. Other files of the workspace are re-validated
against those changes upon a save.
* Fix the ambiguity in the qualified name to use for a Module when
parsing/validating it, now always relying on the Module's qualified name
according to it's environment's resolver (that is, most of the time,
based on the directory names and the file name). However the validation
does not yet reflect this change (as it impacts all existing unit
tests).

Change-Id: I5db29be4ff29090b4ef7137733745fafc881c4ba
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>